+++ docs/src/app/pages/docs/slides/app-slides-concept/app-slides-concept.tsx
@@ -290,14 +290,14 @@ export class AppSlidesConcept {
           </ul>
           <div class="container ion-margin">
             <deckgo-deck embedded={true}>
-              <deckgo-slide-poll poll-link="https://deckdeckgo.com/poll" socket-url="https://api.deckdeckgo.com" connectPollSocket={false}>
+              <deckgo-slide-poll poll-link="https://app.deckdeckgo.com/poll" socket-url="https://api.deckdeckgo.com" connectPollSocket={false}>
                 <h1 slot="question">Do you like my presentation so far?</h1>
                 <p slot="answer-1">It is super</p>
                 <p slot="answer-2">Meh</p>
                 <p slot="answer-3">I could'nt care less</p>
                 <p slot="answer-4">Tell me why</p>
                 <p slot="how-to">
-                  Go to <a href="https://deckdeckgo.com/poll">deckdeckgo.com/poll</a> and use the code {0}
+                  Go to <a href="https://app.deckdeckgo.com/poll">app.deckdeckgo.com/poll</a> and use the code {0}
                 </p>
                 <p slot="awaiting-votes">Awaiting first votes</p>
                 <p slot="answer-5">Ain't nothin' but a heartache</p>