+++ studio/src/app/components/editor/actions/element/app-actions-element/app-actions-element.tsx
@@ -177,7 +177,6 @@ export class AppActionsElement {
       return;
     }
 
-    this.selectedElement.element.removeEventListener('paste', this.cleanOnPaste, true);
     await this.detachMoveToolbarOnElement();
 
     await this.reset();
@@ -209,88 +208,6 @@ export class AppActionsElement {
       const result: HTMLElement = await this.findSelectedElement(element.parentElement);
 
       resolve(result);
-    });
-  }
-
-  private cleanOnPaste = async ($event) => {
-    return new Promise<void>(async (resolve) => {
-      if (!$event || !document) {
-        resolve();
-        return;
-      }
-
-      const parseText: string = $event.clipboardData.getData('text/plain');
-
-      if (!parseText || parseText.length <= 0) {
-        resolve();
-        return;
-      }
-
-      // In the future, Safari isn't yet ready / without preventDefault
-      // // @ts-ignore
-      // navigator.permissions.query({name: 'clipboard-write'}).then(async result => {
-      //     if (result.state === 'granted' || result.state === 'prompt') {
-      //         // @ts-ignore
-      //         await navigator.clipboard.writeText(parseText);
-      //     }
-      // });
-
-      $event.preventDefault();
-
-      const parseTexts: string[] = parseText.replace(/(?:\r\n|\r|\n)/g, '<br/>').split('<br/>');
-
-      if (!parseTexts || parseTexts.length <= 0) {
-        resolve();
-        return;
-      }
-
-      const isCodeElement: boolean = $event.target && $event.target.nodeName && $event.target.nodeName.toLowerCase() === 'code';
-
-      const selected: Selection = await this.getSelection();
-      if (selected && selected.rangeCount) {
-        const range = selected.getRangeAt(0);
-        range.deleteContents();
-
-        parseTexts.forEach((text: string, index: number) => {
-          const newTextNode: Text = document.createTextNode(text);
-          range.collapse(false);
-          range.insertNode(newTextNode);
-
-          if (index < parseTexts.length - 1) {
-            range.collapse(false);
-
-            if (isCodeElement) {
-              const text: Text = document.createTextNode('\n');
-              range.insertNode(text);
-            } else {
-              const br: HTMLBRElement = document.createElement('br');
-              range.insertNode(br);
-            }
-          }
-        });
-
-        selected.empty();
-
-        await this.emitChange();
-      }
-
-      resolve();
-    });
-  };
-
-  private getSelection(): Promise<Selection> {
-    return new Promise<Selection>((resolve) => {
-      let selectedSelection: Selection = null;
-
-      if (window && window.getSelection) {
-        selectedSelection = window.getSelection();
-      } else if (document && document.getSelection) {
-        selectedSelection = document.getSelection();
-      } else if (document && (document as any).selection) {
-        selectedSelection = (document as any).selection.createRange().text;
-      }
-
-      resolve(selectedSelection);
     });
   }
 
@@ -686,8 +603,6 @@ export class AppActionsElement {
       : undefined;
 
     if (element) {
-      element.addEventListener('paste', this.cleanOnPaste, false);
-
       await this.attachResizeSlideContent();
 
       await this.highlightElement(true);