+++ site/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.1 (2020-01-14)
+
+Misc small fix after launch
+
 # 1.0.0 (2020-12-23)
 
 Hello world 👋